Renamed unknown_0348C0 to object_functions
