+++ src/object_functions.c
@@ -1,7 +1,7 @@
 /* The comment below is needed for this file to be picked up by generate_ld */
 /* RAM_POS: 0x80033CC0 */
 
-#include "unknown_0348C0.h"
+#include "object_functions.h"
 
 #include "PR/os_cont.h"
 #include "types.h"
