+++ src/object_functions.h
