+++ src/unknown_043920.c
@@ -14,7 +14,7 @@
 #include "unknown_008C40.h"
 #include "unknown_00BC20.h"
 #include "audio.h"
-#include "unknown_0348C0.h"
+#include "object_functions.h"
 #include "unknown_06B2B0.h"
 
 #define MAX_NUMBER_OF_GHOST_NODES 360
